Move AI related major APIs to Swagger documentation
